app.py
from loader import db
from utils.db_api import db_gino
from utils.send_to_channel import send_first_msg


async def on_startup(dp):
    import filters
    import middlewares
    filters.setup(dp)
    middlewares.setup(dp)

    from utils.notify_admins import on_startup_notify
    print('Connect')
    await db_gino.start_connection(dp)
    await on_startup_notify(dp)


if __name__ == '__main__':
    from aiogram import executor
    from handlers import dp

    executor.start_polling(dp, on_startup=on_startup)

