requirements.txt
requests
aiogram==2.10.1
aiohttp==3.7.2
aiohttp-jinja2==1.3.0
aioredis==1.3.1
async-timeout==3.0.1
asyncio==3.4.3
asyncpg==0.21.0
attrs==20.3.0
Babel==2.8.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
dnspython==1.16.0
eventlet==0.29.1
gino==1.0.1
greenlet==0.4.17
hiredis==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==5.0.0
python-dotenv==0.15.0
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.4
six==1.15.0
SQLAlchemy==1.3.20
typing-extensions==3.7.4.3
Werkzeug==1.0.1
yarl==1.6.2
zope.event==4.5.0
zope.interface==5.2.0
